+++ R/utils.R
@@ -3,3 +3,4 @@ utils::globalVariables(c(".", ":=", "%>%"))
 #' @importFrom dplyr %>%
 #' @export
 dplyr::`%>%`
+